python/agents/genetic-health-coach/pyproject.toml
[project]
name = "genetic-health-coach"
version = "0.1.0"
description = "Agent ADK care generează recomandări personalizate pe baza unui fișier VCF adnotat."
authors = [{ name = "Sample Author", email = "sample-author@google.com" }]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-adk = "^1.0.0"
google-genai = "^1.9.0"
google-cloud-aiplatform = { version = "^1.93.0", extras = ["adk", "agent-engines"] }
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
google-adk = { extras = ["eval"], version = "^1.0.0" }
google-cloud-aiplatform = { version = "^1.93.0", extras = ["adk", "agent-engines", "evaluation"] }
pytest = "^8.3.5"
black = "^25.1.0"
pytest-asyncio = "^0.26.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
